fix: remove model, fix file name, fix return result
